+++ NEWS.md
@@ -3,6 +3,9 @@
 ## argoFloats 0.2.0
 
 * Add subset by section
+* Add "traj" and "bio-traj" arguments to getIndex()
+* Created subtype cycles and trajectories for index type
+* Created map plot for trajectories type
 
 ## argoFloats 0.1.3